Mejorando borrado

+++ Juego/src/MotorGrafico.hpp
@@ -145,7 +145,7 @@ using namespace idsEventos;
             //Cuando muere jugador aparecen 2 botones (Ir a menu o Reiniciar juego)
             void botonesMuerteJugador();
             int getEnemigos_Scena();
-            void EraseColectable(int idx);
+            void EraseColectable(long unsigned int idx);
             void EraseEnemigo(std::size_t i);
             void EraseJugador();
             void EraseArma();
